backend/gaief-function-app/requirements.txt
azure-functions
python-multipart
requests
openai>==1.0.0
azure-ai-formrecognizer